Remove `goerli` from action/script name because target chain can be any chain
